wdk_sys
Module ntddk

Direct FFI bindings to NTDDK APIs from the Windows Driver Kit (WDK)

This module contains all bindings to functions, constants, methods, constructors and destructors in ntddk.h. Types are not included in this module, but are available in the top-level wdk_sys module.

Functions

AcceptSecurityContext⚠

AcquireCredentialsHandleW⚠

AddCredentialsA⚠

AddCredentialsW⚠

ApplyControlToken⚠

BugCheckSecondaryMultiPartDumpDataCallback⚠

CcAsyncCopyRead⚠

CcCanIWrite⚠

CcCoherencyFlushAndPurgeCache⚠

CcCopyRead⚠

CcCopyReadEx⚠

CcCopyWrite⚠

CcCopyWriteEx⚠

CcCopyWriteWontFlush⚠

CcDeferWrite⚠

CcErrorCallbackRoutine⚠

CcFastCopyRead⚠

CcFastCopyWrite⚠

CcFlushCache⚠

CcGetDirtyPages⚠

CcGetFileObjectFromBcb⚠

CcGetFileObjectFromSectionPtrs⚠

CcGetFileObjectFromSectionPtrsRef⚠

CcGetFlushedValidData⚠

CcInitializeCacheMap⚠

CcInitializeCacheMapEx⚠

CcInitializeCacheMapEx2⚠

CcIsCacheManagerCallbackNeeded⚠

CcIsThereDirtyData⚠

CcIsThereDirtyDataEx⚠

CcMapData⚠

CcMdlRead⚠

CcMdlReadComplete⚠

CcMdlWriteAbort⚠

CcMdlWriteComplete⚠

CcPinMappedData⚠

CcPinRead⚠

CcPrepareMdlWrite⚠

CcPreparePinWrite⚠

CcPurgeCacheSection⚠

CcRegisterExternalCacheEx⚠

CcRemapBcb⚠

CcRepinBcb⚠

CcScheduleReadAhead⚠

CcScheduleReadAheadEx⚠

CcSetAdditionalCacheAttributes⚠

CcSetAdditionalCacheAttributesEx⚠

CcSetBcbOwnerPointer⚠

CcSetDirtyPageThreshold⚠

CcSetDirtyPinnedData⚠

CcSetFileSizes⚠

CcSetFileSizesEx⚠

CcSetLogHandleForFile⚠

CcSetParallelFlushFile⚠

CcSetReadAheadGranularity⚠

CcUninitializeCacheMap⚠

CcUnpinData⚠

CcUnpinDataForThread⚠

CcUnpinRepinnedBcb⚠

CcUnregisterExternalCacheEx⚠

CcUpdateExternalCacheInfoEx⚠

CcWaitForCurrentLazyWriterActivity⚠

CcZeroData⚠

ClfsAddLogContainer⚠

ClfsAddLogContainerSet⚠

ClfsAdvanceLogBase⚠

ClfsAlignReservedLog⚠

ClfsAllocReservedLog⚠

ClfsCloseAndResetLogFile⚠

ClfsCloseLogFileObject⚠

ClfsCreateLogFile⚠

ClfsCreateMarshallingArea⚠

ClfsCreateMarshallingAreaEx⚠

ClfsCreateScanContext⚠

ClfsDeleteLogByPointer⚠

ClfsDeleteLogFile⚠

ClfsDeleteMarshallingArea⚠

ClfsEarlierLsn⚠

ClfsFinalize⚠

ClfsFlushBuffers⚠

ClfsFlushToLsn⚠

ClfsFreeReservedLog⚠

ClfsGetContainerName⚠

ClfsGetIoStatistics⚠

ClfsGetLastLsn⚠

ClfsGetLogFileInformation⚠

ClfsInitialize⚠

ClfsLaterLsn⚠

ClfsLsnBlockOffset⚠

ClfsLsnContainer⚠

ClfsLsnCreate⚠

ClfsLsnDifference⚠

ClfsLsnEqual⚠

ClfsLsnGreater⚠

ClfsLsnIncrement⚠

ClfsLsnInvalid⚠

ClfsLsnLess⚠

ClfsLsnNull⚠

ClfsLsnRecordSequence⚠

ClfsMgmtDeregisterManagedClient⚠

ClfsMgmtHandleLogFileFull⚠

ClfsMgmtInstallPolicy⚠

ClfsMgmtQueryPolicy⚠

ClfsMgmtRegisterManagedClient⚠

ClfsMgmtRemovePolicy⚠

ClfsMgmtSetLogFileSize⚠

ClfsMgmtSetLogFileSizeAsClient⚠

ClfsMgmtTailAdvanceFailure⚠

ClfsQueryLogFileInformation⚠

ClfsReadLogRecord⚠

ClfsReadNextLogRecord⚠

ClfsReadPreviousRestartArea⚠

ClfsReadRestartArea⚠

ClfsRemoveLogContainer⚠

ClfsRemoveLogContainerSet⚠

ClfsReserveAndAppendLog⚠

ClfsReserveAndAppendLogAligned⚠

ClfsScanLogContainers⚠

ClfsSetArchiveTail⚠

ClfsSetEndOfLog⚠

ClfsSetLogFileInformation⚠

ClfsTerminateReadLog⚠

ClfsValidTopLevelContext⚠

ClfsWriteRestartArea⚠

CmCallbackGetKeyObjectID⚠

CmCallbackGetKeyObjectIDEx⚠

CmCallbackReleaseKeyObjectIDEx⚠

CmGetBoundTransaction⚠

CmGetCallbackVersion⚠

CmRegisterCallback⚠

CmRegisterCallbackEx⚠

CmSetCallbackObjectContext⚠

CmUnRegisterCallback⚠

CompleteAuthToken⚠

DbgBreakPointWithStatus⚠

DbgPrint⚠

DbgPrintEx⚠

DbgPrintReturnControlC⚠

DbgPrompt⚠

DbgQueryDebugFilterState⚠

DbgSetDebugFilterState⚠

DbgSetDebugPrintCallback⚠

DecryptMessage⚠

DeleteSecurityContext⚠

EncryptMessage⚠

EnumerateSecurityPackagesW⚠

Misc.

EtwActivityIdControl⚠

EtwEventEnabled⚠

EtwProviderEnabled⚠

EtwRegister⚠

EtwSetInformation⚠

EtwUnregister⚠

EtwWrite⚠

EtwWriteEx⚠

EtwWriteString⚠

EtwWriteTransfer⚠

ExAcquireFastMutex⚠

ExAcquireFastMutexUnsafe⚠

ExAcquirePushLockExclusiveEx⚠

ExAcquirePushLockSharedEx⚠

ExAcquireResourceExclusiveLite⚠

ExAcquireResourceSharedLite⚠

ExAcquireRundownProtection⚠

ExAcquireRundownProtectionCacheAware⚠

ExAcquireRundownProtectionCacheAwareEx⚠

ExAcquireRundownProtectionEx⚠

ExAcquireSharedStarveExclusive⚠

ExAcquireSharedWaitForExclusive⚠

ExAcquireSpinLockExclusive⚠

ExAcquireSpinLockExclusiveAtDpcLevel⚠

ExAcquireSpinLockShared⚠

ExAcquireSpinLockSharedAtDpcLevel⚠

ExAdjustLookasideDepth⚠

ExAllocateCacheAwareRundownProtection⚠

ExAllocateFromLookasideListEx⚠

ExAllocateFromNPagedLookasideList⚠

ExAllocateFromPagedLookasideList⚠

ExAllocatePool2⚠

ExAllocatePool3⚠

ExAllocatePoolWithQuota⚠

ExAllocateTimer⚠

ExCancelTimer⚠

ExCleanupRundownProtectionCacheAware⚠

ExConvertExclusiveToSharedLite⚠

ExCreateCallback⚠

ExCreatePool⚠

ExDeleteLookasideListEx⚠

ExDeleteNPagedLookasideList⚠

ExDeletePagedLookasideList⚠

ExDeleteResourceLite⚠

ExDeleteTimer⚠

ExDestroyPool⚠

ExDisableResourceBoostLite⚠

ExEnterCriticalRegionAndAcquireResourceExclusive⚠

ExEnterCriticalRegionAndAcquireResourceShared⚠

ExEnterCriticalRegionAndAcquireSharedWaitForExclusive⚠

ExEnumerateSystemFirmwareTables⚠

ExExtendZone⚠

ExFlushLookasideListEx⚠

ExFreeCacheAwareRundownProtection⚠

ExFreePool⚠

ExFreePool2⚠

ExFreePoolWithTag⚠

ExFreeToLookasideListEx⚠

ExFreeToNPagedLookasideList⚠

ExFreeToPagedLookasideList⚠

ExGetExclusiveWaiterCount⚠

ExGetFirmwareEnvironmentVariable⚠

ExGetFirmwareType⚠

ExGetPreviousMode⚠

ExGetSharedWaiterCount⚠

ExGetSystemFirmwareTable⚠

ExInitializeDeviceAts⚠

ExInitializeLookasideListEx⚠

ExInitializeNPagedLookasideList⚠

ExInitializePagedLookasideList⚠

ExInitializePushLock⚠

ExInitializeResourceLite⚠

ExInitializeRundownProtection⚠

ExInitializeRundownProtectionCacheAware⚠

ExInitializeRundownProtectionCacheAwareEx⚠

ExInitializeZone⚠

ExInterlockedAddLargeInteger⚠

ExInterlockedAddUlong⚠

ExInterlockedExtendZone⚠

ExInterlockedInsertHeadList⚠

ExInterlockedInsertTailList⚠

ExInterlockedPopEntryList⚠

ExInterlockedPushEntryList⚠

ExInterlockedRemoveHeadList⚠

ExIsManufacturingModeEnabled⚠

ExIsProcessorFeaturePresent⚠

ExIsResourceAcquiredExclusiveLite⚠

ExIsResourceAcquiredSharedLite⚠

ExIsSoftBoot⚠

ExLocalTimeToSystemTime⚠

ExNotifyCallback⚠

ExQueryDepthSList⚠

ExQueryPoolBlockSize⚠

ExQueryTimerResolution⚠

ExQueueWorkItem⚠

ExRaiseAccessViolation⚠

ExRaiseDatatypeMisalignment⚠

ExRaiseStatus⚠

ExRcuFreePool⚠

ExReInitializeRundownProtection⚠

ExReInitializeRundownProtectionCacheAware⚠

ExRegisterCallback⚠

ExReinitializeResourceLite⚠

ExReleaseFastMutex⚠

ExReleaseFastMutexUnsafe⚠

ExReleasePushLockExclusiveEx⚠

ExReleasePushLockSharedEx⚠

ExReleaseResourceAndLeaveCriticalRegion⚠

ExReleaseResourceForThreadLite⚠

ExReleaseResourceLite⚠

ExReleaseRundownProtection⚠

ExReleaseRundownProtectionCacheAware⚠

ExReleaseRundownProtectionCacheAwareEx⚠

ExReleaseRundownProtectionEx⚠

ExReleaseSpinLockExclusive⚠

ExReleaseSpinLockExclusiveFromDpcLevel⚠

ExReleaseSpinLockShared⚠

ExReleaseSpinLockSharedFromDpcLevel⚠

ExRundownCompleted⚠

ExRundownCompletedCacheAware⚠

ExSecurePoolUpdate⚠

ExSecurePoolValidate⚠

ExSetFirmwareEnvironmentVariable⚠

ExSetResourceOwnerPointer⚠

ExSetResourceOwnerPointerEx⚠

ExSetTimer⚠

ExSetTimerResolution⚠

ExSizeOfRundownProtectionCacheAware⚠

ExSystemTimeToLocalTime⚠

ExTryAcquirePushLockExclusiveEx⚠

ExTryAcquirePushLockSharedEx⚠

ExTryAcquireSpinLockExclusiveAtDpcLevel⚠

ExTryAcquireSpinLockSharedAtDpcLevel⚠

ExTryConvertSharedSpinLockExclusive⚠

ExTryToAcquireFastMutex⚠

ExUnregisterCallback⚠

ExUuidCreate⚠

ExVerifySuite⚠

ExWaitForRundownProtectionRelease⚠

ExWaitForRundownProtectionReleaseCacheAware⚠

ExpInterlockedFlushSList⚠

ExpInterlockedPopEntrySList⚠

ExpInterlockedPushEntrySList⚠

ExportSecurityContext⚠

Context export/import

FirstEntrySList⚠

FreeContextBuffer⚠

FreeCredentialsHandle⚠

FsRtlAcknowledgeEcp⚠

FsRtlAcquireFileExclusive⚠

FsRtlAddBaseMcbEntry⚠

FsRtlAddBaseMcbEntryEx⚠

FsRtlAddLargeMcbEntry⚠

FsRtlAddMcbEntry⚠

FsRtlAddToTunnelCache⚠

FsRtlAddToTunnelCacheEx⚠

FsRtlAllocateAePushLock⚠

FsRtlAllocateExtraCreateParameter⚠

FsRtlAllocateExtraCreateParameterFromLookasideList⚠

FsRtlAllocateExtraCreateParameterList⚠

FsRtlAllocateFileLock⚠

FsRtlAllocateResource⚠

FsRtlAreNamesEqual⚠

FsRtlAreThereCurrentOrInProgressFileLocks⚠

FsRtlAreThereWaitingFileLocks⚠

FsRtlAreVolumeStartupApplicationsComplete⚠

FsRtlBalanceReads⚠

FsRtlCancellableWaitForMultipleObjects⚠

FsRtlCancellableWaitForSingleObject⚠

FsRtlChangeBackingFileObject⚠

FsRtlCheckFileSystemFilterCallbacksRegistered⚠

FsRtlCheckLockForOplockRequest⚠

FsRtlCheckLockForReadAccess⚠

FsRtlCheckLockForWriteAccess⚠

FsRtlCheckOplock⚠

FsRtlCheckOplockEx⚠

FsRtlCheckOplockEx2⚠

FsRtlCheckOplockForFsFilterCallback⚠

FsRtlCheckUpperOplock⚠

FsRtlCopyRead⚠

FsRtlCopyWrite⚠

FsRtlCreateSectionForDataScan⚠

FsRtlCurrentBatchOplock⚠

FsRtlCurrentOplock⚠

FsRtlCurrentOplockH⚠

FsRtlDeleteExtraCreateParameterLookasideList⚠

FsRtlDeleteKeyFromTunnelCache⚠

FsRtlDeleteTunnelCache⚠

FsRtlDeregisterUncProvider⚠

FsRtlDismountComplete⚠

FsRtlDissectDbcs⚠

FsRtlDissectName⚠

FsRtlDoesDbcsContainWildCards⚠

FsRtlDoesNameContainWildCards⚠

FsRtlFastCheckLockForRead⚠

FsRtlFastCheckLockForWrite⚠

FsRtlFastUnlockAll⚠

FsRtlFastUnlockAllByKey⚠

FsRtlFastUnlockSingle⚠

FsRtlFindExtraCreateParameter⚠

FsRtlFindInTunnelCache⚠

FsRtlFindInTunnelCacheEx⚠

FsRtlFreeAePushLock⚠

FsRtlFreeExtraCreateParameter⚠

FsRtlFreeExtraCreateParameterList⚠

FsRtlFreeFileLock⚠

FsRtlGetCurrentProcessLoaderList⚠

FsRtlGetEcpListFromIrp⚠

FsRtlGetFileSize⚠

FsRtlGetNextBaseMcbEntry⚠

FsRtlGetNextExtraCreateParameter⚠

FsRtlGetNextFileLock⚠

FsRtlGetNextLargeMcbEntry⚠

FsRtlGetNextMcbEntry⚠

FsRtlGetSectorSizeInformation⚠

FsRtlGetSupportedFeatures⚠

FsRtlGetVirtualDiskNestingLevel⚠

FsRtlIncrementCcFastMdlReadWait⚠

FsRtlIncrementCcFastReadNoWait⚠

FsRtlIncrementCcFastReadNotPossible⚠

FsRtlIncrementCcFastReadResourceMiss⚠

FsRtlIncrementCcFastReadWait⚠

FsRtlInitExtraCreateParameterLookasideList⚠

FsRtlInitializeBaseMcb⚠

FsRtlInitializeBaseMcbEx⚠

FsRtlInitializeExtraCreateParameter⚠

FsRtlInitializeExtraCreateParameterList⚠

FsRtlInitializeFileLock⚠

FsRtlInitializeLargeMcb⚠

FsRtlInitializeMcb⚠

FsRtlInitializeOplock⚠

FsRtlInitializeTunnelCache⚠

FsRtlInsertExtraCreateParameter⚠

FsRtlInsertPerFileContext⚠

FsRtlInsertPerFileObjectContext⚠

FsRtlInsertPerStreamContext⚠

FsRtlIs32BitProcess⚠

FsRtlIsDaxVolume⚠

FsRtlIsDbcsInExpression⚠

FsRtlIsEcpAcknowledged⚠

FsRtlIsEcpFromUserMode⚠

FsRtlIsExtentDangling⚠

FsRtlIsFatDbcsLegal⚠

FsRtlIsHpfsDbcsLegal⚠

FsRtlIsMobileOS⚠

FsRtlIsNameInExpression⚠

FsRtlIsNameInUnUpcasedExpression⚠

FsRtlIsNonEmptyDirectoryReparsePointAllowed⚠

FsRtlIsNtstatusExpected⚠

FsRtlIsPagingFile⚠

FsRtlIsSystemPagingFile⚠

FsRtlIsTotalDeviceFailure⚠

FsRtlIssueDeviceIoControl⚠

FsRtlKernelFsControlFile⚠

FsRtlLogCcFlushError⚠

FsRtlLookupBaseMcbEntry⚠

FsRtlLookupLargeMcbEntry⚠

FsRtlLookupLastBaseMcbEntry⚠

FsRtlLookupLastBaseMcbEntryAndIndex⚠

FsRtlLookupLastLargeMcbEntry⚠

FsRtlLookupLastLargeMcbEntryAndIndex⚠

FsRtlLookupLastMcbEntry⚠

FsRtlLookupMcbEntry⚠

FsRtlLookupPerFileContext⚠

FsRtlLookupPerFileObjectContext⚠

FsRtlLookupPerStreamContextInternal⚠

FsRtlMdlReadCompleteDev⚠

FsRtlMdlReadDev⚠

FsRtlMdlReadEx⚠

FsRtlMdlWriteCompleteDev⚠

FsRtlMupGetProviderIdFromName⚠

FsRtlMupGetProviderInfoFromFileObject⚠

FsRtlNormalizeNtstatus⚠

FsRtlNotifyCleanup⚠

FsRtlNotifyCleanupAll⚠

FsRtlNotifyFilterChangeDirectory⚠

FsRtlNotifyFilterReportChange⚠

FsRtlNotifyFullChangeDirectory⚠

FsRtlNotifyFullReportChange⚠

FsRtlNotifyInitializeSync⚠

FsRtlNotifyStreamFileObject⚠

FsRtlNotifyUninitializeSync⚠

FsRtlNotifyVolumeEvent⚠

FsRtlNotifyVolumeEventEx⚠

FsRtlNumberOfRunsInBaseMcb⚠

FsRtlNumberOfRunsInLargeMcb⚠

FsRtlNumberOfRunsInMcb⚠

FsRtlOplockBreakH⚠

FsRtlOplockBreakH2⚠

FsRtlOplockBreakToNone⚠

FsRtlOplockBreakToNoneEx⚠

FsRtlOplockFsctrl⚠

FsRtlOplockFsctrlEx⚠

FsRtlOplockGetAnyBreakOwnerProcess⚠

FsRtlOplockIsFastIoPossible⚠

FsRtlOplockIsSharedRequest⚠

FsRtlOplockKeysEqual⚠

FsRtlPostPagingFileStackOverflow⚠

FsRtlPostStackOverflow⚠

FsRtlPrepareMdlWriteDev⚠

FsRtlPrepareMdlWriteEx⚠

FsRtlPrepareToReuseEcp⚠

FsRtlPrivateLock⚠

FsRtlProcessFileLock⚠

FsRtlQueryCachedVdl⚠

FsRtlQueryInformationFile⚠

FsRtlQueryKernelEaFile⚠

FsRtlQueryMaximumVirtualDiskNestingLevel⚠

FsRtlRegisterFileSystemFilterCallbacks⚠

FsRtlRegisterUncProvider⚠

FsRtlRegisterUncProviderEx⚠

FsRtlRegisterUncProviderEx2⚠

FsRtlReleaseFile⚠

FsRtlRemoveBaseMcbEntry⚠

FsRtlRemoveDotsFromPath⚠

FsRtlRemoveExtraCreateParameter⚠

FsRtlRemoveLargeMcbEntry⚠

FsRtlRemoveMcbEntry⚠

FsRtlRemovePerFileContext⚠

FsRtlRemovePerFileObjectContext⚠

FsRtlRemovePerStreamContext⚠

FsRtlResetBaseMcb⚠

FsRtlResetLargeMcb⚠

FsRtlSetDriverBacking⚠

FsRtlSetEcpListIntoIrp⚠

FsRtlSetKernelEaFile⚠

FsRtlSplitBaseMcb⚠

FsRtlSplitLargeMcb⚠

FsRtlTeardownPerFileContexts⚠

FsRtlTeardownPerStreamContexts⚠

FsRtlTruncateBaseMcb⚠

FsRtlTruncateLargeMcb⚠

FsRtlTruncateMcb⚠

FsRtlUninitializeBaseMcb⚠

FsRtlUninitializeFileLock⚠

FsRtlUninitializeLargeMcb⚠

FsRtlUninitializeMcb⚠

FsRtlUninitializeOplock⚠

FsRtlUpdateDiskCounters⚠

FsRtlUpperOplockFsctrl⚠

FsRtlValidateReparsePointBuffer⚠

FsRtlVolumeDeviceToCorrelationId⚠

GetSecurityUserInfo⚠

HalAcquireDisplayOwnership⚠

HalAllocateCrashDumpRegisters⚠

HalAllocateHardwareCounters⚠

HalBugCheckSystem⚠

HalDmaAllocateCrashDumpRegistersEx⚠

HalDmaFreeCrashDumpRegistersEx⚠

HalExamineMBR⚠

HalFreeHardwareCounters⚠

HalGetAdapter⚠

HalGetBusDataByOffset⚠

HalGetScatterGatherList⚠

HalPutDmaAdapter⚠

HalPutScatterGatherList⚠

HalSetBusDataByOffset⚠

HalTranslateBusAddress⚠

HvlRegisterWheaErrorNotification⚠

HvlUnregisterWheaErrorNotification⚠

ImpersonateSecurityContext⚠

ImportSecurityContextW⚠

InitSecurityInterfaceW⚠

InitializeSListHead⚠

InitializeSecurityContextW⚠

Context Management Functions

InterlockedPushListSList⚠

IoAcquireCancelSpinLock⚠

IoAcquireKsrPersistentMemory⚠

IoAcquireKsrPersistentMemoryEx⚠

IoAcquireRemoveLockEx⚠

IoAcquireVpbSpinLock⚠

IoAllocateController⚠

IoAllocateDriverObjectExtension⚠

IoAllocateErrorLogEntry⚠

IoAllocateIrp⚠

IoAllocateIrpEx⚠

IoAllocateMdl⚠

IoAllocateSfioStreamIdentifier⚠

IoAllocateWorkItem⚠

IoApplyPriorityInfoThread⚠

IoAssignResources⚠

IoAttachDevice⚠

IoAttachDeviceByPointer⚠

IoAttachDeviceToDeviceStack⚠

IoAttachDeviceToDeviceStackSafe⚠

IoBuildAsynchronousFsdRequest⚠

IoBuildDeviceIoControlRequest⚠

IoBuildPartialMdl⚠

IoBuildSynchronousFsdRequest⚠

IoCancelFileOpen⚠

IoCancelIrp⚠

IoCheckDesiredAccess⚠

IoCheckEaBufferValidity⚠

IoCheckFileObjectOpenedAsCopyDestination⚠

IoCheckFileObjectOpenedAsCopySource⚠

IoCheckFunctionAccess⚠

IoCheckLinkShareAccess⚠

IoCheckQuerySetFileInformation⚠

IoCheckQuerySetVolumeInformation⚠

IoCheckQuotaBufferValidity⚠

IoCheckShareAccess⚠

IoCheckShareAccessEx⚠

IoCleanupIrp⚠

IoClearActivityIdThread⚠

IoClearFsTrackOffsetState⚠

IoClearIrpExtraCreateParameter⚠

IoConnectInterrupt⚠

IoConnectInterruptEx⚠

IoCreateController⚠

IoCreateDevice⚠

IoCreateDisk⚠

IoCreateDriverProxyExtension⚠

IoCreateFile⚠

IoCreateFileEx⚠

IoCreateFileSpecifyDeviceObjectHint⚠

IoCreateNotificationEvent⚠

IoCreateStreamFileObject⚠

IoCreateStreamFileObjectEx⚠

IoCreateStreamFileObjectEx2⚠

IoCreateStreamFileObjectLite⚠

IoCreateSymbolicLink⚠

IoCreateSynchronizationEvent⚠

IoCreateSystemThread⚠

IoCreateUnprotectedSymbolicLink⚠

IoCsqInitialize⚠

IoCsqInitializeEx⚠

IoCsqInsertIrp⚠

IoCsqInsertIrpEx⚠

IoCsqRemoveIrp⚠

IoCsqRemoveNextIrp⚠

IoDecrementKeepAliveCount⚠

IoDeleteController⚠

IoDeleteDevice⚠

IoDeleteSymbolicLink⚠

IoDetachDevice⚠

IoDisconnectInterrupt⚠

IoDisconnectInterruptEx⚠

IoDriverProxyCreateHotSwappableWorkerThread⚠

IoEnumerateDeviceObjectList⚠

IoEnumerateKsrPersistentMemoryEx⚠

IoEnumerateRegisteredFiltersList⚠

IoFastQueryNetworkAttributes⚠

IoForwardIrpSynchronously⚠

IoFreeController⚠

IoFreeErrorLogEntry⚠

IoFreeIrp⚠

IoFreeKsrPersistentMemory⚠

IoFreeMdl⚠

IoFreeSfioStreamIdentifier⚠

IoFreeWorkItem⚠

IoGetActivityIdIrp⚠

IoGetActivityIdThread⚠

IoGetAffinityInterrupt⚠

IoGetAttachedDevice⚠

IoGetAttachedDeviceReference⚠

IoGetBaseFileSystemDeviceObject⚠

IoGetBootDiskInformation⚠

IoGetBootDiskInformationLite⚠

IoGetConfigurationInformation⚠

IoGetContainerInformation⚠

IoGetCopyInformationExtension⚠

IoGetCurrentProcess⚠

IoGetDeviceAttachmentBaseRef⚠

IoGetDeviceDirectory⚠

IoGetDeviceInterfaceAlias⚠

IoGetDeviceInterfacePropertyData⚠

IoGetDeviceInterfaces⚠

IoGetDeviceNumaNode⚠

IoGetDeviceObjectPointer⚠

IoGetDeviceProperty⚠

IoGetDevicePropertyData⚠

IoGetDeviceToVerify⚠

IoGetDiskDeviceObject⚠

IoGetDmaAdapter⚠

IoGetDriverDirectory⚠

IoGetDriverObjectExtension⚠

IoGetDriverProxyEndpointWrapper⚠

IoGetDriverProxyExtensionFromDriverObject⚠

IoGetDriverProxyFeatures⚠

IoGetFileObjectGenericMapping⚠

IoGetFsTrackOffsetState⚠

IoGetFsZeroingOffset⚠

IoGetInitialStack⚠

IoGetInitiatorProcess⚠

IoGetIoAttributionHandle⚠

IoGetIoPriorityHint⚠

IoGetIommuInterface⚠

IoGetIommuInterfaceEx⚠

IoGetIrpExtraCreateParameter⚠

IoGetKsrPersistentMemoryBuffer⚠

IoGetLowerDeviceObject⚠

IoGetOplockKeyContext⚠

IoGetOplockKeyContextEx⚠

IoGetPagingIoPriority⚠

IoGetRelatedDeviceObject⚠

IoGetRequestorProcess⚠

IoGetRequestorProcessId⚠

IoGetRequestorSessionId⚠

IoGetSfioStreamIdentifier⚠

IoGetShadowFileInformation⚠

IoGetSilo⚠

IoGetSiloParameters⚠

IoGetStackLimits⚠

IoGetTopLevelIrp⚠

IoGetTransactionParameterBlock⚠

IoIncrementKeepAliveCount⚠

IoInitializeIrp⚠

IoInitializeIrpEx⚠

IoInitializeRemoveLockEx⚠

IoInitializeTimer⚠

IoInitializeWorkItem⚠

IoInvalidateDeviceRelations⚠

IoInvalidateDeviceState⚠

IoIrpHasFsTrackOffsetExtensionType⚠

IoIs32bitProcess⚠

IoIsFileObjectIgnoringSharing⚠

IoIsFileOriginRemote⚠

IoIsInitiator32bitProcess⚠

IoIsOperationSynchronous⚠

IoIsSystemThread⚠

IoIsValidIrpStatus⚠

IoIsValidNameGraftingBuffer⚠

IoIsWdmVersionAvailable⚠

IoMakeAssociatedIrp⚠

IoMakeAssociatedIrpEx⚠

IoMapKsrPersistentMemoryEx⚠

IoOpenDeviceInterfaceRegistryKey⚠

IoOpenDeviceRegistryKey⚠

IoOpenDriverRegistryKey⚠

IoPageRead⚠

IoPropagateActivityIdToThread⚠

IoQueryDeviceDescription⚠

IoQueryFileDosDeviceName⚠

IoQueryFileInformation⚠

IoQueryFullDriverPath⚠

IoQueryInformationByName⚠

IoQueryKsrPersistentMemorySize⚠

IoQueryKsrPersistentMemorySizeEx⚠

IoQueryVolumeInformation⚠

IoQueueThreadIrp⚠

IoQueueWorkItem⚠

IoQueueWorkItemEx⚠

IoRaiseHardError⚠

IoRaiseInformationalHardError⚠

IoReadDiskSignature⚠

IoReadPartitionTable⚠

IoReadPartitionTableEx⚠

IoRecordIoAttribution⚠

IoRegisterBootDriverCallback⚠

IoRegisterBootDriverReinitialization⚠

IoRegisterContainerNotification⚠

IoRegisterDeviceInterface⚠

IoRegisterDriverProxyEndpoints⚠

IoRegisterDriverReinitialization⚠

IoRegisterFileSystem⚠

IoRegisterFsRegistrationChange⚠

IoRegisterFsRegistrationChangeMountAware⚠

IoRegisterLastChanceShutdownNotification⚠

IoRegisterPlugPlayNotification⚠

IoRegisterShutdownNotification⚠

IoReleaseCancelSpinLock⚠

IoReleaseRemoveLockAndWaitEx⚠

IoReleaseRemoveLockEx⚠

IoReleaseVpbSpinLock⚠

IoRemoveLinkShareAccess⚠

IoRemoveLinkShareAccessEx⚠

IoRemoveShareAccess⚠

IoReplaceFileObjectName⚠

IoReplacePartitionUnit⚠

IoReportDetectedDevice⚠

IoReportInterruptActive⚠

IoReportInterruptInactive⚠

IoReportResourceForDetection⚠

IoReportResourceUsage⚠

IoReportRootDevice⚠

IoReportTargetDeviceChange⚠

IoReportTargetDeviceChangeAsynchronous⚠

IoRequestDeviceEject⚠

IoRequestDeviceEjectEx⚠

IoRequestDeviceRemovalForReset⚠

IoReserveKsrPersistentMemory⚠

IoReserveKsrPersistentMemoryEx⚠

IoRetrievePriorityInfo⚠

IoReuseIrp⚠

IoSetActivityIdIrp⚠

IoSetActivityIdThread⚠

IoSetCompletionRoutineEx⚠

IoSetDeviceInterfacePropertyData⚠

IoSetDeviceInterfaceState⚠

IoSetDevicePropertyData⚠

IoSetDeviceToVerify⚠

IoSetFileObjectIgnoreSharing⚠

IoSetFileOrigin⚠

IoSetFsTrackOffsetState⚠

IoSetFsZeroingOffset⚠

IoSetFsZeroingOffsetRequired⚠

IoSetHardErrorOrVerifyDevice⚠

IoSetInformation⚠

IoSetIoAttributionIrp⚠

IoSetIoPriorityHint⚠

IoSetIrpExtraCreateParameter⚠

IoSetLinkShareAccess⚠

IoSetMasterIrpStatus⚠

IoSetPartitionInformation⚠

IoSetPartitionInformationEx⚠

IoSetShadowFileInformation⚠

IoSetShareAccess⚠

IoSetShareAccessEx⚠

IoSetStartIoAttributes⚠

IoSetSystemPartition⚠

IoSetThreadHardErrorMode⚠

IoSetTopLevelIrp⚠

IoSizeOfIrpEx⚠

IoSizeofWorkItem⚠

IoStartNextPacket⚠

IoStartNextPacketByKey⚠

IoStartPacket⚠

IoStartTimer⚠

IoStopTimer⚠

IoSynchronousCallDriver⚠

IoSynchronousPageWrite⚠

IoThreadToProcess⚠

IoTransferActivityId⚠

IoTranslateBusAddress⚠

IoTryQueueWorkItem⚠

IoUninitializeWorkItem⚠

IoUnregisterBootDriverCallback⚠

IoUnregisterContainerNotification⚠

IoUnregisterFileSystem⚠

IoUnregisterFsRegistrationChange⚠

IoUnregisterPlugPlayNotification⚠

IoUnregisterPlugPlayNotificationEx⚠

IoUnregisterShutdownNotification⚠

IoUpdateLinkShareAccess⚠

IoUpdateLinkShareAccessEx⚠

IoUpdateShareAccess⚠

IoValidateDeviceIoControlAccess⚠

IoVerifyPartitionTable⚠

IoVerifyVolume⚠

IoVolumeDeviceNameToGuid⚠

IoVolumeDeviceNameToGuidPath⚠

IoVolumeDeviceToDosName⚠

IoVolumeDeviceToGuid⚠

IoVolumeDeviceToGuidPath⚠

IoWMIAllocateInstanceIds⚠

IoWMIDeviceObjectToInstanceName⚠

IoWMIDeviceObjectToProviderId⚠

IoWMIExecuteMethod⚠

IoWMIHandleToInstanceName⚠

IoWMIOpenBlock⚠

IoWMIQueryAllData⚠

IoWMIQueryAllDataMultiple⚠

IoWMIQuerySingleInstance⚠

IoWMIQuerySingleInstanceMultiple⚠

IoWMIRegistrationControl⚠

IoWMISetNotificationCallback⚠

IoWMISetSingleInstance⚠

IoWMISetSingleItem⚠

IoWMISuggestInstanceName⚠

IoWMIWriteEvent⚠

IoWithinStackLimits⚠

IoWriteErrorLogEntry⚠

IoWriteKsrPersistentMemory⚠

IoWritePartitionTable⚠

IoWritePartitionTableEx⚠

IofCallDriver⚠

IofCompleteRequest⚠

IofGetDriverProxyWrapperFromEndpoint⚠

KdChangeOption⚠

KdDisableDebugger⚠

KdEnableDebugger⚠

KdRefreshDebuggerNotPresent⚠

KeAcquireGuardedMutex⚠

KeAcquireGuardedMutexUnsafe⚠

KeAcquireInStackQueuedSpinLock⚠

KeAcquireInStackQueuedSpinLockAtDpcLevel⚠

KeAcquireInStackQueuedSpinLockForDpc⚠

KeAcquireInterruptSpinLock⚠

KeAcquireQueuedSpinLock⚠

KeAcquireSpinLockAtDpcLevel⚠

KeAcquireSpinLockForDpc⚠

KeAcquireSpinLockRaiseToDpc⚠

KeAcquireSpinLockRaiseToSynch⚠

KeAddTriageDumpDataBlock⚠

KeAreAllApcsDisabled⚠

KeAreApcsDisabled⚠

KeAttachProcess⚠

KeBugCheck⚠

KeBugCheckEx⚠

KeCancelTimer⚠

KeClearEvent⚠

KeConvertAuxiliaryCounterToPerformanceCounter⚠

KeConvertPerformanceCounterToAuxiliaryCounter⚠

KeDelayExecutionThread⚠

KeDeregisterBoundCallback⚠

KeDeregisterBugCheckCallback⚠

KeDeregisterBugCheckReasonCallback⚠

KeDeregisterNmiCallback⚠

KeDeregisterProcessorChangeCallback⚠

KeDetachProcess⚠

KeEnterCriticalRegion⚠

KeEnterGuardedRegion⚠

KeExpandKernelStackAndCallout⚠

KeExpandKernelStackAndCalloutEx⚠

KeFlushIoBuffers⚠

KeFlushQueuedDpcs⚠

KeFlushWriteBuffer⚠

KeGetCurrentIrql⚠

KeGetCurrentNodeNumber⚠

KeGetCurrentProcessorNumberEx⚠

KeGetProcessorIndexFromNumber⚠

KeGetProcessorNumberFromIndex⚠

KeGetRecommendedSharedDataAlignment⚠

KeInitializeCrashDumpHeader⚠

KeInitializeDeviceQueue⚠

KeInitializeDpc⚠

KeInitializeEvent⚠

KeInitializeGuardedMutex⚠

KeInitializeMutant⚠

KeInitializeMutex⚠

KeInitializeQueue⚠

KeInitializeSemaphore⚠

KeInitializeSpinLock⚠

KeInitializeThreadedDpc⚠

KeInitializeTimer⚠

KeInitializeTimerEx⚠

KeInitializeTriageDumpDataArray⚠

KeInsertByKeyDeviceQueue⚠

KeInsertDeviceQueue⚠

KeInsertHeadQueue⚠

KeInsertQueue⚠

KeInsertQueueDpc⚠

KeInvalidateAllCaches⚠

KeInvalidateRangeAllCaches⚠

KeIpiGenericCall⚠

KeIsExecutingDpc⚠

KeLeaveCriticalRegion⚠

KeLeaveGuardedRegion⚠

KeLowerIrql⚠

KePulseEvent⚠

KeQueryActiveGroupCount⚠

KeQueryActiveProcessorCount⚠

KeQueryActiveProcessorCountEx⚠

KeQueryActiveProcessors⚠

KeQueryAuxiliaryCounterFrequency⚠

KeQueryDpcWatchdogInformation⚠

KeQueryGroupAffinity⚠

KeQueryHardwareCounterConfiguration⚠

KeQueryHighestNodeNumber⚠

KeQueryInterruptTimePrecise⚠

KeQueryLogicalProcessorRelationship⚠

KeQueryMaximumGroupCount⚠

KeQueryMaximumProcessorCount⚠

KeQueryMaximumProcessorCountEx⚠

KeQueryNodeActiveAffinity⚠

KeQueryNodeActiveAffinity2⚠

KeQueryNodeActiveProcessorCount⚠

KeQueryNodeMaximumProcessorCount⚠

KeQueryOwnerMutant⚠

KeQueryPerformanceCounter⚠

KeQueryPriorityThread⚠

KeQueryRuntimeThread⚠

KeQuerySystemTimePrecise⚠

KeQueryTimeIncrement⚠

KeQueryTotalCycleTimeThread⚠

KeQueryUnbiasedInterruptTime⚠

KeQueryUnbiasedInterruptTimePrecise⚠

KeRcuReadLock⚠

KeRcuReadUnlock⚠

KeRcuSynchronize⚠

KeReadStateEvent⚠

KeReadStateMutant⚠

KeReadStateMutex⚠

KeReadStateQueue⚠

KeReadStateSemaphore⚠

KeReadStateTimer⚠

KeRegisterBoundCallback⚠

KeRegisterBugCheckCallback⚠

KeRegisterBugCheckReasonCallback⚠

KeRegisterNmiCallback⚠

KeRegisterProcessorChangeCallback⚠

KeReleaseGuardedMutex⚠

KeReleaseGuardedMutexUnsafe⚠

KeReleaseInStackQueuedSpinLock⚠

KeReleaseInStackQueuedSpinLockForDpc⚠

KeReleaseInStackQueuedSpinLockFromDpcLevel⚠

KeReleaseInterruptSpinLock⚠

KeReleaseMutant⚠

KeReleaseMutex⚠

KeReleaseQueuedSpinLock⚠

KeReleaseSemaphore⚠

KeReleaseSpinLock⚠

KeReleaseSpinLockForDpc⚠

KeReleaseSpinLockFromDpcLevel⚠

KeRemoveByKeyDeviceQueue⚠

KeRemoveByKeyDeviceQueueIfBusy⚠

KeRemoveDeviceQueue⚠

KeRemoveEntryDeviceQueue⚠

KeRemoveQueue⚠

KeRemoveQueueDpc⚠

KeRemoveQueueDpcEx⚠

KeRemoveQueueEx⚠

KeResetEvent⚠

KeRestoreExtendedProcessorState⚠

KeRevertToUserAffinityThread⚠

KeRevertToUserAffinityThreadEx⚠

KeRevertToUserGroupAffinityThread⚠

KeRundownQueue⚠

KeSaveExtendedProcessorState⚠

KeSetBasePriorityThread⚠

KeSetCoalescableTimer⚠

KeSetEvent⚠

KeSetHardwareCounterConfiguration⚠

KeSetIdealProcessorThread⚠

KeSetImportanceDpc⚠

KeSetKernelStackSwapEnable⚠

KeSetPriorityThread⚠

KeSetSystemAffinityThread⚠

KeSetSystemAffinityThreadEx⚠

KeSetSystemGroupAffinityThread⚠

KeSetTargetProcessorDpc⚠

KeSetTargetProcessorDpcEx⚠

KeSetTimer⚠

KeSetTimerEx⚠

KeShouldYieldProcessor⚠

KeSrcuAllocate⚠

KeSrcuFree⚠

KeSrcuReadLock⚠

KeSrcuReadUnlock⚠

KeSrcuSynchronize⚠

KeStackAttachProcess⚠

KeStallExecutionProcessor⚠

KeSynchronizeExecution⚠

KeTestSpinLock⚠

KeTryToAcquireGuardedMutex⚠

KeTryToAcquireQueuedSpinLock⚠

KeTryToAcquireSpinLockAtDpcLevel⚠

KeUnstackDetachProcess⚠

KeWaitForMultipleObjects⚠

KeWaitForSingleObject⚠

KfRaiseIrql⚠

LsaFreeReturnBuffer⚠

LsaLogonUser⚠

LsaRegisterLogonProcess⚠

MakeSignature⚠

Message Support API

MapSecurityError⚠

MmAddPhysicalMemory⚠

MmAddVerifierSpecialThunks⚠

MmAddVerifierThunks⚠

MmAdvanceMdl⚠

MmAllocateContiguousMemory⚠

MmAllocateContiguousMemoryEx⚠

MmAllocateContiguousMemorySpecifyCache⚠

MmAllocateContiguousMemorySpecifyCacheNode⚠

MmAllocateContiguousNodeMemory⚠

MmAllocateMappingAddress⚠

MmAllocateMappingAddressEx⚠

MmAllocateMdlForIoSpace⚠

MmAllocateNodePagesForMdlEx⚠

MmAllocateNonCachedMemory⚠

MmAllocatePagesForMdl⚠

MmAllocatePagesForMdlEx⚠

MmAllocatePartitionNodePagesForMdlEx⚠

MmAreMdlPagesCached⚠

MmBuildMdlForNonPagedPool⚠

MmCanFileBeTruncated⚠

MmCopyMemory⚠

MmCreateMdl⚠

MmCreateMirror⚠

MmDoesFileHaveUserWritableReferences⚠

MmFlushImageSection⚠

MmForceSectionClosed⚠

MmForceSectionClosedEx⚠

MmFreeContiguousMemory⚠

MmFreeContiguousMemorySpecifyCache⚠

MmFreeMappingAddress⚠

MmFreeNonCachedMemory⚠

MmFreePagesFromMdl⚠

MmFreePagesFromMdlEx⚠

MmGetCacheAttribute⚠

MmGetCacheAttributeEx⚠

MmGetMaximumFileSectionSize⚠

MmGetPhysicalAddress⚠

MmGetPhysicalMemoryRanges⚠

MmGetPhysicalMemoryRangesEx⚠

MmGetPhysicalMemoryRangesEx2⚠

MmGetSystemRoutineAddress⚠

MmGetVirtualForPhysical⚠

MmIsAddressValid⚠

MmIsDriverSuspectForVerifier⚠

MmIsDriverVerifying⚠

MmIsDriverVerifyingByAddress⚠

MmIsFileSectionActive⚠

MmIsIoSpaceActive⚠

MmIsNonPagedSystemAddressValid⚠

MmIsRecursiveIoFault⚠

MmIsThisAnNtAsSystem⚠

MmIsVerifierEnabled⚠

MmLockPagableDataSection⚠

MmLockPagableSectionByHandle⚠

MmMapIoSpace⚠

MmMapIoSpaceEx⚠

MmMapLockedPages⚠

MmMapLockedPagesSpecifyCache⚠

MmMapLockedPagesWithReservedMapping⚠

MmMapMdl⚠

MmMapMemoryDumpMdlEx⚠

MmMapUserAddressesToPage⚠

MmMapVideoDisplay⚠

MmMapViewInSessionSpace⚠

MmMapViewInSessionSpaceEx⚠

MmMapViewInSystemSpace⚠

MmMapViewInSystemSpaceEx⚠

MmMdlPageContentsState⚠

MmMdlPagesAreZero⚠

MmPageEntireDriver⚠

MmPrefetchPages⚠

MmProbeAndLockPages⚠

MmProbeAndLockProcessPages⚠

MmProbeAndLockSelectedPages⚠

MmProtectDriverSection⚠

MmProtectMdlSystemAddress⚠

MmQuerySystemSize⚠

MmRemovePhysicalMemory⚠

MmResetDriverPaging⚠

MmRotatePhysicalView⚠

MmSecureVirtualMemory⚠

MmSecureVirtualMemoryEx⚠

MmSetAddressRangeModified⚠

MmSetPermanentCacheAttribute⚠

MmSizeOfMdl⚠

MmUnlockPagableImageSection⚠

MmUnlockPages⚠

MmUnmapIoSpace⚠

MmUnmapLockedPages⚠

MmUnmapReservedMapping⚠

MmUnmapVideoDisplay⚠

MmUnmapViewInSessionSpace⚠

MmUnmapViewInSystemSpace⚠

MmUnsecureVirtualMemory⚠

NtAccessCheckAndAuditAlarm⚠

NtAccessCheckByTypeAndAuditAlarm⚠

NtAccessCheckByTypeResultListAndAuditAlarm⚠

NtAccessCheckByTypeResultListAndAuditAlarmByHandle⚠

NtAdjustGroupsToken⚠

NtAdjustPrivilegesToken⚠

NtAllocateVirtualMemory⚠

NtClose⚠

NtCloseObjectAuditAlarm⚠

NtCommitComplete⚠

NtCommitEnlistment⚠

NtCommitTransaction⚠

NtCopyFileChunk⚠

NtCreateEnlistment⚠

NtCreateFile⚠

NtCreateResourceManager⚠

NtCreateSection⚠

NtCreateSectionEx⚠

NtCreateTransaction⚠

NtCreateTransactionManager⚠

NtDeleteObjectAuditAlarm⚠

NtDeviceIoControlFile⚠

NtDuplicateToken⚠

NtEnumerateTransactionObject⚠

NtFilterToken⚠

NtFlushBuffersFileEx⚠

NtFreeVirtualMemory⚠

NtFsControlFile⚠

NtGetNotificationResourceManager⚠

NtImpersonateAnonymousToken⚠

NtLockFile⚠

NtManagePartition⚠

NtOpenEnlistment⚠

NtOpenFile⚠

NtOpenJobObjectToken⚠

NtOpenObjectAuditAlarm⚠

NtOpenProcess⚠

NtOpenProcessToken⚠

NtOpenProcessTokenEx⚠

NtOpenRegistryTransaction⚠

NtOpenResourceManager⚠

NtOpenThreadToken⚠

NtOpenThreadTokenEx⚠

NtOpenTransaction⚠

NtOpenTransactionManager⚠

NtPowerInformation⚠

NtPrePrepareComplete⚠

NtPrePrepareEnlistment⚠

NtPrepareComplete⚠

NtPrepareEnlistment⚠

NtPrivilegeCheck⚠

NtPrivilegeObjectAuditAlarm⚠

NtPrivilegedServiceAuditAlarm⚠

NtPropagationComplete⚠

NtPropagationFailed⚠

NtQueryDirectoryFile⚠

NtQueryDirectoryFileEx⚠

NtQueryInformationByName⚠

NtQueryInformationEnlistment⚠

NtQueryInformationFile⚠

NtQueryInformationResourceManager⚠

NtQueryInformationToken⚠

NtQueryInformationTransaction⚠

NtQueryInformationTransactionManager⚠

NtQueryObject⚠

NtQueryQuotaInformationFile⚠

NtQuerySecurityObject⚠

NtQueryVirtualMemory⚠

NtQueryVolumeInformationFile⚠

NtReadFile⚠

NtReadOnlyEnlistment⚠

NtRecoverEnlistment⚠

NtRecoverResourceManager⚠

NtRecoverTransactionManager⚠

NtRegisterProtocolAddressInformation⚠

NtRenameTransactionManager⚠

NtRollbackComplete⚠

NtRollbackEnlistment⚠

NtRollbackRegistryTransaction⚠

NtRollbackTransaction⚠

NtRollforwardTransactionManager⚠

NtSetInformationEnlistment⚠

NtSetInformationFile⚠

NtSetInformationResourceManager⚠

NtSetInformationThread⚠

NtSetInformationToken⚠

NtSetInformationTransaction⚠

NtSetInformationTransactionManager⚠

NtSetInformationVirtualMemory⚠

NtSetQuotaInformationFile⚠

NtSetSecurityObject⚠

NtSetVolumeInformationFile⚠

NtSinglePhaseReject⚠

NtUnlockFile⚠

NtWriteFile⚠

ObCloseHandle⚠

ObDereferenceObjectDeferDelete⚠

ObDereferenceObjectDeferDeleteWithTag⚠

ObGetFilterVersion⚠

ObGetObjectSecurity⚠

ObInsertObject⚠

ObIsKernelHandle⚠

ObMakeTemporaryObject⚠

ObOpenObjectByPointer⚠

ObOpenObjectByPointerWithTag⚠

ObQueryNameString⚠

ObQueryObjectAuditingByHandle⚠

ObReferenceObjectByHandle⚠

ObReferenceObjectByHandleWithTag⚠

ObReferenceObjectByPointer⚠

ObReferenceObjectByPointerWithTag⚠

ObReferenceObjectSafe⚠

ObReferenceObjectSafeWithTag⚠

ObRegisterCallbacks⚠

ObReleaseObjectSecurity⚠

ObUnRegisterCallbacks⚠

ObfDereferenceObject⚠

ObfDereferenceObjectWithTag⚠

ObfReferenceObject⚠

ObfReferenceObjectWithTag⚠

PcwAddInstance⚠

PcwCloseInstance⚠

PcwCreateInstance⚠

PcwRegister⚠

PcwUnregister⚠

PfxFindPrefix⚠

PfxInitialize⚠

PfxInsertPrefix⚠

PfxRemovePrefix⚠

PoCallDriver⚠

PoClearPowerRequest⚠

PoCreatePowerLimitRequest⚠

PoCreatePowerRequest⚠

PoCreateThermalRequest⚠

PoDeletePowerLimitRequest⚠

PoDeletePowerRequest⚠

PoDeleteThermalRequest⚠

PoEndDeviceBusy⚠

PoFxActivateComponent⚠

PoFxCompleteDevicePowerNotRequired⚠

PoFxCompleteDirectedPowerDown⚠

PoFxCompleteIdleCondition⚠

PoFxCompleteIdleState⚠

PoFxIdleComponent⚠

PoFxIssueComponentPerfStateChange⚠

PoFxIssueComponentPerfStateChangeMultiple⚠

PoFxNotifySurprisePowerOn⚠

PoFxPowerControl⚠

PoFxPowerOnCrashdumpDevice⚠

PoFxQueryCurrentComponentPerfState⚠

PoFxRegisterComponentPerfStates⚠

PoFxRegisterCrashdumpDevice⚠

PoFxRegisterDevice⚠

PoFxRegisterDripsWatchdogCallback⚠

PoFxReportDevicePoweredOn⚠

PoFxSetComponentLatency⚠

PoFxSetComponentResidency⚠

PoFxSetComponentWake⚠

PoFxSetDeviceIdleTimeout⚠

PoFxSetTargetDripsDevicePowerState⚠

PoFxStartDevicePowerManagement⚠

PoFxUnregisterDevice⚠

PoGetSystemWake⚠

PoGetThermalRequestSupport⚠

PoQueryPowerLimitAttributes⚠

PoQueryPowerLimitValue⚠

PoQueryWatchdogTime⚠

PoQueueShutdownWorkItem⚠

PoRegisterDeviceForIdleDetection⚠

PoRegisterForEffectivePowerModeNotifications⚠

PoRegisterPowerSettingCallback⚠

PoRegisterSystemState⚠

PoRequestPowerIrp⚠

PoSetDeviceBusyEx⚠

PoSetHiberRange⚠

PoSetPowerLimitValue⚠

PoSetPowerRequest⚠

PoSetPowerState⚠

PoSetSystemState⚠

PoSetSystemWake⚠

PoSetSystemWakeDevice⚠

PoSetThermalActiveCooling⚠

PoSetThermalPassiveCooling⚠

PoStartDeviceBusy⚠

PoStartNextPowerIrp⚠

PoUnregisterFromEffectivePowerModeNotifications⚠

PoUnregisterPowerSettingCallback⚠

PoUnregisterSystemState⚠

ProbeForRead⚠

ProbeForWrite⚠

PsAcquireSiloHardReference⚠

PsAllocSiloContextSlot⚠

PsAllocateAffinityToken⚠

PsAssignImpersonationToken⚠

PsAttachSiloToCurrentThread⚠

PsChargePoolQuota⚠

PsChargeProcessPoolQuota⚠

PsCreateSiloContext⚠

PsCreateSystemThread⚠

PsDereferenceImpersonationToken⚠

PsDereferencePrimaryToken⚠

PsDereferenceSiloContext⚠

PsDetachSiloFromCurrentThread⚠

PsDisableImpersonation⚠

PsFreeAffinityToken⚠

PsFreeSiloContextSlot⚠

PsGetCurrentProcessId⚠

PsGetCurrentServerSilo⚠

PsGetCurrentServerSiloName⚠

PsGetCurrentSilo⚠

PsGetCurrentThreadId⚠

PsGetCurrentThreadTeb⚠

PsGetEffectiveServerSilo⚠

PsGetHostSilo⚠

PsGetJobServerSilo⚠

PsGetJobSilo⚠

PsGetParentSilo⚠

PsGetPermanentSiloContext⚠

PsGetProcessCreateTimeQuadPart⚠

PsGetProcessExitStatus⚠

PsGetProcessExitTime⚠

PsGetProcessId⚠

PsGetProcessStartKey⚠

PsGetServerSiloActiveConsoleId⚠

PsGetServerSiloServiceSessionId⚠

PsGetSiloContainerId⚠

PsGetSiloContext⚠

PsGetSiloMonitorContextSlot⚠

PsGetThreadCreateTime⚠

PsGetThreadExitStatus⚠

PsGetThreadId⚠

PsGetThreadProcess⚠

PsGetThreadProcessId⚠

PsGetThreadProperty⚠

PsGetThreadServerSilo⚠

PsGetVersion⚠

PsImpersonateClient⚠

PsInsertPermanentSiloContext⚠

PsInsertSiloContext⚠

PsIsCurrentThreadInServerSilo⚠

PsIsCurrentThreadPrefetching⚠

PsIsDiskCountersEnabled⚠

PsIsHostSilo⚠

PsIsSystemThread⚠

PsIsThreadTerminating⚠

PsLookupProcessByProcessId⚠

PsLookupThreadByThreadId⚠

PsMakeSiloContextPermanent⚠

PsQueryProcessAvailableCpus⚠

PsQueryProcessAvailableCpusCount⚠

PsQuerySystemAvailableCpus⚠

PsQuerySystemAvailableCpusCount⚠

PsQueryTotalCycleTimeProcess⚠

PsReferenceImpersonationToken⚠

PsReferencePrimaryToken⚠

PsReferenceSiloContext⚠

PsRegisterProcessAvailableCpusChangeNotification⚠

PsRegisterSiloMonitor⚠

PsRegisterSystemAvailableCpusChangeNotification⚠

PsReleaseSiloHardReference⚠

PsRemoveCreateThreadNotifyRoutine⚠

PsRemoveLoadImageNotifyRoutine⚠

PsRemoveSiloContext⚠

PsReplaceSiloContext⚠

PsRestoreImpersonation⚠

PsReturnPoolQuota⚠

PsRevertToSelf⚠

PsRevertToUserMultipleGroupAffinityThread⚠

PsSetCreateProcessNotifyRoutine⚠

PsSetCreateProcessNotifyRoutineEx⚠

PsSetCreateProcessNotifyRoutineEx2⚠

PsSetCreateThreadNotifyRoutine⚠

PsSetCreateThreadNotifyRoutineEx⚠

PsSetCurrentThreadPrefetching⚠

PsSetLoadImageNotifyRoutine⚠

PsSetLoadImageNotifyRoutineEx⚠

PsSetSystemMultipleGroupAffinityThread⚠

PsStartSiloMonitor⚠

PsTerminateServerSilo⚠

PsTerminateSystemThread⚠

PsUnregisterAvailableCpusChangeNotification⚠

PsUnregisterSiloMonitor⚠

PsUpdateDiskCounters⚠

PsWrapApcWow64Thread⚠

PshedAllocateMemory⚠

PshedFreeMemory⚠

PshedIsSystemWheaEnabled⚠

PshedRegisterPlugin⚠

PshedSynchronizeExecution⚠

PshedUnregisterPlugin⚠

QueryContextAttributesExW⚠

QueryContextAttributesW⚠

QueryCredentialsAttributesExW⚠

QueryCredentialsAttributesW⚠

QuerySecurityContextToken⚠

QuerySecurityPackageInfoW⚠

RevertSecurityContext⚠

RtlAbsoluteToSelfRelativeSD⚠

RtlAddAccessAllowedAce⚠

RtlAddAccessAllowedAceEx⚠

RtlAddAce⚠

RtlAllocateAndInitializeSid⚠

RtlAllocateAndInitializeSidEx⚠

RtlAllocateHeap⚠

RtlAnsiStringToUnicodeString⚠

RtlAppendStringToString⚠

RtlAppendUnicodeStringToString⚠

RtlAppendUnicodeToString⚠

RtlAreBitsClear⚠

RtlAreBitsSet⚠

RtlAssert⚠

RtlCaptureContext⚠

RtlCaptureContext2⚠

RtlCaptureStackBackTrace⚠

RtlCharToInteger⚠

RtlCheckRegistryKey⚠

RtlClearAllBits⚠

RtlClearBit⚠

RtlClearBits⚠

RtlCmDecodeMemIoResource⚠

RtlCmEncodeMemIoResource⚠

RtlCompareAltitudes⚠

RtlCompareMemory⚠

RtlCompareMemoryUlong⚠

RtlCompareString⚠

RtlCompareUnicodeString⚠

RtlCompareUnicodeStrings⚠

RtlCompressBuffer⚠

RtlCompressChunks⚠

RtlContractHashTable⚠

RtlConvertSidToUnicodeString⚠

RtlCopyBitMap⚠

RtlCopyDeviceMemory⚠

RtlCopyLuid⚠

RtlCopyMemoryNonTemporal⚠

RtlCopySid⚠

RtlCopyString⚠

RtlCopyUnicodeString⚠

RtlCopyVolatileMemory⚠

RtlCrc32⚠

RtlCrc64⚠

RtlCreateAcl⚠

RtlCreateHashTable⚠

RtlCreateHashTableEx⚠

RtlCreateHeap⚠

RtlCreateRegistryKey⚠

RtlCreateSecurityDescriptor⚠

RtlCreateSecurityDescriptorRelative⚠

RtlCreateServiceSid⚠

RtlCreateSystemVolumeInformationFolder⚠

RtlCreateUnicodeString⚠

RtlCreateVirtualAccountSid⚠

RtlCustomCPToUnicodeN⚠

RtlDecompressBuffer⚠

RtlDecompressBufferEx⚠

RtlDecompressBufferEx2⚠

RtlDecompressChunks⚠

RtlDecompressFragment⚠

RtlDecompressFragmentEx⚠

RtlDelete⚠

RtlDeleteAce⚠

RtlDeleteElementGenericTable⚠

RtlDeleteElementGenericTableAvl⚠

RtlDeleteElementGenericTableAvlEx⚠

RtlDeleteHashTable⚠

RtlDeleteNoSplay⚠

RtlDeleteRegistryValue⚠

RtlDescribeChunk⚠

RtlDestroyHeap⚠

RtlDowncaseUnicodeChar⚠

RtlDowncaseUnicodeString⚠

RtlDrainNonVolatileFlush⚠

RtlDuplicateUnicodeString⚠

RtlEndEnumerationHashTable⚠

RtlEndStrongEnumerationHashTable⚠

RtlEndWeakEnumerationHashTable⚠

RtlEnumerateEntryHashTable⚠

RtlEnumerateGenericTable⚠

RtlEnumerateGenericTableAvl⚠

RtlEnumerateGenericTableLikeADirectory⚠

RtlEnumerateGenericTableWithoutSplaying⚠

RtlEnumerateGenericTableWithoutSplayingAvl⚠

RtlEqualPrefixSid⚠

RtlEqualSid⚠

RtlEqualString⚠

RtlEqualUnicodeString⚠

RtlExpandHashTable⚠

RtlExtendCorrelationVector⚠

RtlExtractBitMap⚠

RtlFillMemoryNonTemporal⚠

RtlFillNonVolatileMemory⚠

RtlFindClearBits⚠

RtlFindClearBitsAndSet⚠

RtlFindClearRuns⚠

RtlFindClosestEncodableLength⚠

RtlFindFirstRunClear⚠

RtlFindLastBackwardRunClear⚠

RtlFindLeastSignificantBit⚠

RtlFindLongestRunClear⚠

RtlFindMostSignificantBit⚠

RtlFindNextForwardRunClear⚠

RtlFindSetBits⚠

RtlFindSetBitsAndClear⚠

RtlFindUnicodePrefix⚠

RtlFlushNonVolatileMemory⚠

RtlFlushNonVolatileMemoryRanges⚠

RtlFreeAnsiString⚠

RtlFreeHeap⚠

RtlFreeNonVolatileToken⚠

RtlFreeOemString⚠

RtlFreeSid⚠

RtlFreeUTF8String⚠

RtlFreeUnicodeString⚠

RtlGUIDFromString⚠

RtlGenerate8dot3Name⚠

RtlGenerateClass5Guid⚠

RtlGetAce⚠

RtlGetAcesBufferSize⚠

RtlGetActiveConsoleId⚠

RtlGetCallersAddress⚠

RtlGetCompressionWorkSpaceSize⚠

RtlGetConsoleSessionForegroundProcessId⚠

RtlGetDaclSecurityDescriptor⚠

RtlGetElementGenericTable⚠

RtlGetElementGenericTableAvl⚠

RtlGetEnabledExtendedFeatures⚠

RtlGetGroupSecurityDescriptor⚠

RtlGetNextEntryHashTable⚠

RtlGetNonVolatileToken⚠

RtlGetNtProductType⚠

RtlGetNtSystemRoot⚠

RtlGetOwnerSecurityDescriptor⚠

RtlGetPersistedStateLocation⚠

RtlGetProductInfo⚠

RtlGetSaclSecurityDescriptor⚠

RtlGetSuiteMask⚠

RtlGetSystemGlobalData⚠

RtlGetVersion⚠

RtlHashUnicodeString⚠

RtlIdentifierAuthoritySid⚠

RtlIdnToAscii⚠

RtlIdnToNameprepUnicode⚠

RtlIdnToUnicode⚠

RtlIncrementCorrelationVector⚠

RtlInitAnsiString⚠

RtlInitAnsiStringEx⚠

RtlInitCodePageTable⚠

RtlInitEnumerationHashTable⚠

RtlInitString⚠

RtlInitStringEx⚠

RtlInitStrongEnumerationHashTable⚠

RtlInitUTF8String⚠

RtlInitUTF8StringEx⚠

RtlInitUnicodeString⚠

RtlInitUnicodeStringEx⚠

RtlInitWeakEnumerationHashTable⚠

RtlInitializeBitMap⚠

RtlInitializeCorrelationVector⚠

RtlInitializeGenericTable⚠

RtlInitializeGenericTableAvl⚠

RtlInitializeSid⚠

RtlInitializeSidEx⚠

RtlInitializeUnicodePrefix⚠

RtlInsertElementGenericTable⚠

RtlInsertElementGenericTableAvl⚠

RtlInsertElementGenericTableFull⚠

RtlInsertElementGenericTableFullAvl⚠

RtlInsertEntryHashTable⚠

RtlInsertUnicodePrefix⚠

RtlInt64ToUnicodeString⚠

RtlIntegerToUnicodeString⚠

RtlIoDecodeMemIoResource⚠

RtlIoEncodeMemIoResource⚠

RtlIsApiSetImplemented⚠

RtlIsCloudFilesPlaceholder⚠

RtlIsGenericTableEmpty⚠

RtlIsGenericTableEmptyAvl⚠

RtlIsMultiSessionSku⚠

RtlIsMultiUsersInSessionSku⚠

RtlIsNameLegalDOS8Dot3⚠

RtlIsNonEmptyDirectoryReparsePointAllowed⚠

RtlIsNormalizedString⚠

RtlIsNtDdiVersionAvailable⚠

RtlIsPartialPlaceholder⚠

RtlIsPartialPlaceholderFileHandle⚠

RtlIsPartialPlaceholderFileInfo⚠

RtlIsSandboxedToken⚠

RtlIsSandboxedTokenHandle⚠

RtlIsServicePackVersionInstalled⚠

RtlIsStateSeparationEnabled⚠

RtlIsUntrustedObject⚠

RtlIsValidOemCharacter⚠

RtlIsZeroMemory⚠

RtlLengthRequiredSid⚠

RtlLengthSecurityDescriptor⚠

RtlLengthSid⚠

RtlLookupElementGenericTable⚠

RtlLookupElementGenericTableAvl⚠

RtlLookupElementGenericTableFull⚠

RtlLookupElementGenericTableFullAvl⚠

RtlLookupEntryHashTable⚠

RtlLookupFirstMatchingElementGenericTableAvl⚠

RtlMapGenericMask⚠

RtlMoveVolatileMemory⚠

RtlMultiByteToUnicodeN⚠

RtlMultiByteToUnicodeSize⚠

RtlNextUnicodePrefix⚠

RtlNormalizeSecurityDescriptor⚠

RtlNormalizeString⚠

RtlNtStatusToDosError⚠

RtlNtStatusToDosErrorNoTeb⚠

RtlNumberGenericTableElements⚠

RtlNumberGenericTableElementsAvl⚠

RtlNumberOfClearBits⚠

RtlNumberOfClearBitsInRange⚠

RtlNumberOfSetBits⚠

RtlNumberOfSetBitsInRange⚠

RtlNumberOfSetBitsUlongPtr⚠

RtlOemStringToCountedUnicodeString⚠

RtlOemStringToUnicodeString⚠

RtlOemToUnicodeN⚠

RtlOsDeploymentState⚠

RtlPrefetchMemoryNonTemporal⚠

RtlPrefixString⚠

RtlPrefixUnicodeString⚠

RtlQueryPackageIdentity⚠

RtlQueryPackageIdentityEx⚠

RtlQueryProcessPlaceholderCompatibilityMode⚠

RtlQueryRegistryValueWithFallback⚠

RtlQueryRegistryValues⚠

RtlQueryThreadPlaceholderCompatibilityMode⚠

RtlQueryValidationRunlevel⚠

RtlRaiseCustomSystemEventTrigger⚠

RtlRandom⚠

RtlRandomEx⚠

RtlRealPredecessor⚠

RtlRealSuccessor⚠

RtlRemoveEntryHashTable⚠

RtlRemoveUnicodePrefix⚠

RtlReplaceSidInSd⚠

RtlReserveChunk⚠

RtlRunOnceBeginInitialize⚠

RtlRunOnceComplete⚠

RtlRunOnceExecuteOnce⚠

RtlRunOnceInitialize⚠

RtlSecondsSince1970ToTime⚠

RtlSecondsSince1980ToTime⚠

RtlSelfRelativeToAbsoluteSD⚠

RtlSetAllBits⚠

RtlSetBit⚠

RtlSetBits⚠

RtlSetDaclSecurityDescriptor⚠

RtlSetGroupSecurityDescriptor⚠

RtlSetOwnerSecurityDescriptor⚠

RtlSetProcessPlaceholderCompatibilityMode⚠

RtlSetSystemGlobalData⚠

RtlSetThreadPlaceholderCompatibilityMode⚠

RtlSetVolatileMemory⚠

RtlSplay⚠

RtlStringFromGUID⚠

RtlStronglyEnumerateEntryHashTable⚠

RtlSubAuthorityCountSid⚠

RtlSubAuthoritySid⚠

RtlSubtreePredecessor⚠

RtlSubtreeSuccessor⚠

RtlSuffixUnicodeString⚠

RtlTestBit⚠

RtlTimeFieldsToTime⚠

RtlTimeToSecondsSince1970⚠

RtlTimeToSecondsSince1980⚠

RtlTimeToTimeFields⚠

RtlUTF8StringToUnicodeString⚠

RtlUTF8ToUnicodeN⚠

RtlUnicodeStringToAnsiString⚠

RtlUnicodeStringToCountedOemString⚠

RtlUnicodeStringToInt64⚠

RtlUnicodeStringToInteger⚠

RtlUnicodeStringToOemString⚠

RtlUnicodeStringToUTF8String⚠

RtlUnicodeToCustomCPN⚠

RtlUnicodeToMultiByteN⚠

RtlUnicodeToMultiByteSize⚠

RtlUnicodeToOemN⚠

RtlUnicodeToUTF8N⚠

RtlUpcaseUnicodeChar⚠

RtlUpcaseUnicodeString⚠

RtlUpcaseUnicodeStringToCountedOemString⚠

RtlUpcaseUnicodeStringToOemString⚠

RtlUpcaseUnicodeToCustomCPN⚠

RtlUpcaseUnicodeToMultiByteN⚠

RtlUpcaseUnicodeToOemN⚠

RtlUpperChar⚠

RtlUpperString⚠

RtlValidRelativeSecurityDescriptor⚠

RtlValidSecurityDescriptor⚠

RtlValidSid⚠

RtlValidateCorrelationVector⚠

RtlValidateUnicodeString⚠

RtlVerifyVersionInfo⚠

RtlVolumeDeviceToDosName⚠

RtlWalkFrameChain⚠

RtlWeaklyEnumerateEntryHashTable⚠

RtlWriteNonVolatileMemory⚠

RtlWriteRegistryValue⚠

RtlxAnsiStringToUnicodeSize⚠

RtlxOemStringToUnicodeSize⚠

RtlxUnicodeStringToAnsiSize⚠

RtlxUnicodeStringToOemSize⚠

SeAccessCheck⚠

SeAccessCheckFromState⚠

SeAccessCheckFromStateEx⚠

SeAdjustAccessStateForAccessConstraints⚠

SeAdjustAccessStateForTrustLabel⚠

SeAdjustObjectSecurity⚠

SeAppendPrivileges⚠

SeAssignSecurity⚠

SeAssignSecurityEx⚠

SeAuditFipsCryptoSelftests⚠

SeAuditHardLinkCreation⚠

SeAuditHardLinkCreationWithTransaction⚠

SeAuditTransactionStateChange⚠

SeAuditingAnyFileEventsWithContext⚠

SeAuditingAnyFileEventsWithContextEx⚠

SeAuditingFileEvents⚠

SeAuditingFileEventsWithContext⚠

SeAuditingFileEventsWithContextEx⚠

SeAuditingFileOrGlobalEvents⚠

SeAuditingHardLinkEvents⚠

SeAuditingHardLinkEventsWithContext⚠

SeCaptureSubjectContext⚠

SeCaptureSubjectContextEx⚠

SeCheckForCriticalAceRemoval⚠

SeComputeAutoInheritByObjectType⚠

SeCreateClientSecurity⚠

SeCreateClientSecurityFromSubjectContext⚠

SeDeassignSecurity⚠

SeDeleteClientSecurity⚠

SeDeleteObjectAuditAlarm⚠

SeDeleteObjectAuditAlarmWithTransaction⚠

SeEtwWriteKMCveEvent⚠

SeExamineGlobalSacl⚠

SeExamineSacl⚠

SeFilterToken⚠

SeFreePrivileges⚠

SeImpersonateClient⚠

SeImpersonateClientEx⚠

SeLocateProcessImageName⚠

SeLockSubjectContext⚠

SeMarkLogonSessionForTerminationNotification⚠

SeMarkLogonSessionForTerminationNotificationEx⚠

SeMaximumAuditMaskFromGlobalSacl⚠

SeObjectCreateSaclAccessBits⚠

SeOpenObjectAuditAlarm⚠

SeOpenObjectAuditAlarmWithTransaction⚠

SeOpenObjectForDeleteAuditAlarm⚠

SeOpenObjectForDeleteAuditAlarmWithTransaction⚠

SePrivilegeCheck⚠

SeQueryAuthenticationIdToken⚠

SeQueryInformationToken⚠

SeQuerySecurityDescriptorInfo⚠

SeQueryServerSiloToken⚠

SeQuerySessionIdToken⚠

SeQuerySessionIdTokenEx⚠

SeQueryTokenIntegrity⚠

SeRegisterImageVerificationCallback⚠

SeRegisterLogonSessionTerminatedRoutine⚠

SeRegisterLogonSessionTerminatedRoutineEx⚠

SeReleaseSubjectContext⚠

SeReportSecurityEvent⚠

SeReportSecurityEventWithSubCategory⚠

SeSetAccessStateGenericMapping⚠

SeSetAuditParameter⚠

SeSetSecurityDescriptorInfo⚠

SeSetSecurityDescriptorInfoEx⚠

SeSetSessionIdToken⚠

SeShouldCheckForAccessRightsFromParent⚠

SeSinglePrivilegeCheck⚠

SeTokenFromAccessInformation⚠

SeTokenGetNoChildProcessRestricted⚠

SeTokenGetRedirectionTrustPolicy⚠

SeTokenIsAdmin⚠

SeTokenIsNoChildProcessRestrictionEnforced⚠

SeTokenIsRestricted⚠

SeTokenIsWriteRestricted⚠

SeTokenSetNoChildProcessRestricted⚠

SeTokenSetRedirectionTrustPolicy⚠

SeTokenType⚠

SeUnlockSubjectContext⚠

SeUnregisterImageVerificationCallback⚠

SeUnregisterLogonSessionTerminatedRoutine⚠

SeUnregisterLogonSessionTerminatedRoutineEx⚠

SeValidSecurityDescriptor⚠

SecAllocateAndSetCallTarget⚠

SecAllocateAndSetIPAddress⚠

SecFreeCallContext⚠

SecLookupAccountName⚠

SecLookupAccountSid⚠

SecLookupWellKnownSid⚠

SecMakeSPN⚠

SecMakeSPNEx⚠

SecMakeSPNEx2⚠

SetContextAttributesW⚠

SetCredentialsAttributesW⚠

SspiAcceptSecurityContextAsync⚠

SspiAcquireCredentialsHandleAsyncA⚠

SspiAcquireCredentialsHandleAsyncW⚠

SspiAsyncContextRequiresNotify⚠

SspiCompareAuthIdentities⚠

SspiCopyAuthIdentity⚠

SspiCreateAsyncContext⚠

SspiDeleteSecurityContextAsync⚠

SspiEncodeAuthIdentityAsStrings⚠

SspiEncodeStringsAsAuthIdentity⚠

SspiExcludePackage⚠

SspiFreeAsyncContext⚠

SspiFreeAuthIdentity⚠

SspiFreeCredentialsHandleAsync⚠

SspiGetAsyncCallStatus⚠

SspiGetTargetHostName⚠

SspiInitializeSecurityContextAsyncA⚠

SspiInitializeSecurityContextAsyncW⚠

SspiIsPromptingNeeded⚠

SspiLocalFree⚠

SspiMarshalAuthIdentity⚠

SspiReinitAsyncContext⚠

SspiSetAsyncNotifyCallback⚠

SspiUnmarshalAuthIdentity⚠

SspiValidateAuthIdentity⚠

SspiZeroAuthIdentity⚠

TmCommitComplete⚠

TmCommitEnlistment⚠

TmCommitTransaction⚠

TmCreateEnlistment⚠

TmDereferenceEnlistmentKey⚠

TmEnableCallbacks⚠

TmGetTransactionId⚠

TmInitializeTransactionManager⚠

TmIsTransactionActive⚠

TmPrePrepareComplete⚠

TmPrePrepareEnlistment⚠

TmPrepareComplete⚠

TmPrepareEnlistment⚠

TmPropagationComplete⚠

TmPropagationFailed⚠

TmReadOnlyEnlistment⚠

TmRecoverEnlistment⚠

TmRecoverResourceManager⚠

TmRecoverTransactionManager⚠

TmReferenceEnlistmentKey⚠

TmRenameTransactionManager⚠

TmRequestOutcomeEnlistment⚠

TmRollbackComplete⚠

TmRollbackEnlistment⚠

TmRollbackTransaction⚠

TmSinglePhaseReject⚠

UnsignedMultiply128⚠

VerSetConditionMask⚠

VerifySignature⚠

VslCreateSecureSection⚠

VslDeleteSecureSection⚠

WheaAddErrorSource⚠

WheaAddErrorSourceDeviceDriver⚠

WheaAddErrorSourceDeviceDriverV1⚠

WheaAddHwErrorReportSectionDeviceDriver⚠

WheaAttemptRowOffline⚠

WheaConfigureErrorSource⚠

WheaCreateHwErrorReportDeviceDriver⚠

WheaErrorSourceGetState⚠

WheaGetErrorSourceInfo⚠

WheaGetNotifyAllOfflinesPolicy⚠

WheaHighIrqlLogSelEventHandlerRegister⚠

WheaHighIrqlLogSelEventHandlerUnregister⚠

WheaHwErrorReportAbandonDeviceDriver⚠

WheaHwErrorReportSetSectionNameDeviceDriver⚠

WheaHwErrorReportSetSeverityDeviceDriver⚠

WheaHwErrorReportSubmitDeviceDriver⚠

WheaInitializeRecordHeader⚠

WheaIsCriticalState⚠

WheaIsLogSelHandlerInitialized⚠

WheaLogInternalEvent⚠

WheaPrmTranslateDimmAddress⚠

WheaPrmTranslatePhysicalAddress⚠

WheaRegisterErrorSourceOverride⚠

WheaRegisterInUsePageOfflineNotification⚠

WheaRemoveErrorSource⚠

WheaRemoveErrorSourceDeviceDriver⚠

WheaReportHwError⚠

WheaReportHwErrorDeviceDriver⚠

WheaSignalHandlerOverride⚠

WheaUnconfigureErrorSource⚠

WheaUnregisterErrorSourceOverride⚠

WheaUnregisterInUsePageOfflineNotification⚠

WmiQueryTraceInformation⚠

ZwAllocateLocallyUniqueId⚠

ZwAllocateVirtualMemory⚠

ZwAllocateVirtualMemoryEx⚠

ZwCancelTimer⚠

ZwClose⚠

ZwCommitComplete⚠

ZwCommitEnlistment⚠

ZwCommitRegistryTransaction⚠

ZwCommitTransaction⚠

ZwCopyFileChunk⚠

ZwCreateDirectoryObject⚠

ZwCreateEnlistment⚠

ZwCreateEvent⚠

ZwCreateFile⚠

ZwCreateKey⚠

ZwCreateKeyTransacted⚠

ZwCreateRegistryTransaction⚠

ZwCreateResourceManager⚠

ZwCreateSection⚠

ZwCreateTimer⚠

ZwCreateTransaction⚠

ZwCreateTransactionManager⚠

ZwDeleteFile⚠

ZwDeleteKey⚠

ZwDeleteValueKey⚠

ZwDeviceIoControlFile⚠

ZwDisplayString⚠

ZwDuplicateObject⚠

ZwDuplicateToken⚠

ZwEnumerateKey⚠

ZwEnumerateTransactionObject⚠

ZwEnumerateValueKey⚠

ZwFlushBuffersFile⚠

ZwFlushBuffersFileEx⚠

ZwFlushKey⚠

ZwFlushVirtualMemory⚠

ZwFreeVirtualMemory⚠

ZwFsControlFile⚠

ZwGetNotificationResourceManager⚠

ZwLoadDriver⚠

ZwLockFile⚠

ZwMakeTemporaryObject⚠

ZwMapViewOfSection⚠

ZwMapViewOfSectionEx⚠

ZwNotifyChangeKey⚠

ZwOpenDirectoryObject⚠

ZwOpenEnlistment⚠

ZwOpenEvent⚠

ZwOpenFile⚠

ZwOpenKey⚠

ZwOpenKeyEx⚠

ZwOpenKeyTransacted⚠

ZwOpenKeyTransactedEx⚠

ZwOpenProcess⚠

ZwOpenProcessTokenEx⚠

ZwOpenResourceManager⚠

ZwOpenSection⚠

ZwOpenSymbolicLinkObject⚠

ZwOpenThreadTokenEx⚠

ZwOpenTimer⚠

ZwOpenTransaction⚠

ZwOpenTransactionManager⚠

ZwPowerInformation⚠

ZwPrePrepareComplete⚠

ZwPrePrepareEnlistment⚠

ZwPrepareComplete⚠

ZwPrepareEnlistment⚠

ZwQueryDirectoryFile⚠

ZwQueryDirectoryFileEx⚠

ZwQueryEaFile⚠

ZwQueryFullAttributesFile⚠

ZwQueryInformationByName⚠

ZwQueryInformationEnlistment⚠

ZwQueryInformationFile⚠

ZwQueryInformationResourceManager⚠

ZwQueryInformationToken⚠

ZwQueryInformationTransaction⚠

ZwQueryInformationTransactionManager⚠

ZwQueryKey⚠

ZwQueryObject⚠

ZwQueryQuotaInformationFile⚠

ZwQuerySecurityObject⚠

ZwQuerySymbolicLinkObject⚠

ZwQueryValueKey⚠

ZwQueryVirtualMemory⚠

ZwQueryVolumeInformationFile⚠

ZwReadFile⚠

ZwReadOnlyEnlistment⚠

ZwRecoverEnlistment⚠

ZwRecoverResourceManager⚠

ZwRecoverTransactionManager⚠

ZwRenameKey⚠

ZwRestoreKey⚠

ZwRollbackComplete⚠

ZwRollbackEnlistment⚠

ZwRollbackTransaction⚠

ZwRollforwardTransactionManager⚠

ZwSaveKey⚠

ZwSaveKeyEx⚠

ZwSetEaFile⚠

ZwSetEvent⚠

ZwSetInformationEnlistment⚠

ZwSetInformationFile⚠

ZwSetInformationKey⚠

ZwSetInformationResourceManager⚠

ZwSetInformationThread⚠

ZwSetInformationToken⚠

ZwSetInformationTransaction⚠

ZwSetInformationTransactionManager⚠

ZwSetInformationVirtualMemory⚠

ZwSetQuotaInformationFile⚠

ZwSetSecurityObject⚠

ZwSetTimer⚠

ZwSetTimerEx⚠

ZwSetValueKey⚠

ZwSetVolumeInformationFile⚠

ZwSinglePhaseReject⚠

ZwTerminateProcess⚠

ZwUnloadDriver⚠

ZwUnlockFile⚠

ZwUnmapViewOfSection⚠

ZwWaitForSingleObject⚠

ZwWriteFile⚠

_BitScanForward⚠

_BitScanForward64⚠

_BitScanReverse⚠

_BitScanReverse64⚠

_InterlockedAnd⚠

_InterlockedAnd8⚠

_InterlockedAnd16⚠

_InterlockedAnd64⚠

_InterlockedCompareExchange⚠

_InterlockedCompareExchange16⚠

_InterlockedCompareExchange64⚠

_InterlockedCompareExchange128⚠

_InterlockedCompareExchangePointer⚠

_InterlockedDecrement⚠

_InterlockedDecrement16⚠

_InterlockedDecrement64⚠

_InterlockedExchange⚠

_InterlockedExchange8⚠

_InterlockedExchange16⚠

_InterlockedExchange64⚠

_InterlockedExchangeAdd⚠

_InterlockedExchangeAdd8⚠

_InterlockedExchangeAdd64⚠

_InterlockedExchangePointer⚠

_InterlockedIncrement⚠

_InterlockedIncrement16⚠

_InterlockedIncrement64⚠

_InterlockedOr⚠

_InterlockedOr8⚠

_InterlockedOr16⚠

_InterlockedOr64⚠

_InterlockedXor⚠

_InterlockedXor8⚠

_InterlockedXor16⚠

_InterlockedXor64⚠

_RTL_CONSTANT_STRING_type_check⚠

_ReadWriteBarrier⚠

_ReturnAddress⚠

__C_specific_handler⚠

__addgsbyte⚠

__addgsdword⚠

__addgsqword⚠

__addgsword⚠

__cpuidex⚠

__fastfail⚠

__faststorefence⚠

__getcallerseflags⚠

__inbyte⚠

__inbytestring⚠

__incgsbyte⚠

__incgsdword⚠

__incgsqword⚠

__incgsword⚠

__indword⚠

__indwordstring⚠

__int2c⚠

__inword⚠

__inwordstring⚠

__isascii⚠

__iscsym⚠

__iscsymf⚠

__iswcsym⚠

__iswcsymf⚠

__movsb⚠

__movsd⚠

__movsq⚠

__movsw⚠

__mulh⚠

__outbyte⚠

__outbytestring⚠

__outdword⚠

__outdwordstring⚠

__outword⚠

__outwordstring⚠

__popcnt64⚠

__rdtsc⚠

__readcr8⚠

__readgsbyte⚠

__readgsdword⚠

__readgsqword⚠

__readgsword⚠

__readpmc⚠

__segmentlimit⚠

__shiftleft128⚠

__shiftright128⚠

__stosb⚠

__stosd⚠

__stosq⚠

__stosw⚠

__toascii⚠

__umulh⚠

__va_start⚠

__wcserror⚠

__wcserror_s⚠

__writecr8⚠

__writegsbyte⚠

__writegsdword⚠

__writegsqword⚠

__writegsword⚠

_abnormal_termination⚠

_bittest⚠

_bittest64⚠

_bittestandcomplement⚠

_bittestandcomplement64⚠

_bittestandreset⚠

_bittestandreset64⚠

_bittestandset⚠

_bittestandset64⚠

_byteswap_uint64⚠

_byteswap_ulong⚠

_byteswap_ushort⚠

_exception_code⚠

_exception_info⚠

_guard_check_icall⚠

_guard_icall_checks_enforced⚠

_guard_rf_checks_enforced⚠

_interlockedbittestandreset⚠

_interlockedbittestandreset64⚠

_interlockedbittestandset⚠

_interlockedbittestandset64⚠

_isalnum_l⚠

_isalpha_l⚠

_iscntrl_l⚠

_isctype⚠

_isctype_l⚠

_isdigit_l⚠

_isgraph_l⚠

_isleadbyte_l⚠

_islower_l⚠

_isprint_l⚠

_ispunct_l⚠

_isspace_l⚠

_isupper_l⚠

_iswalnum_l⚠

_iswalpha_l⚠

_iswcntrl_l⚠

_iswcsym_l⚠

_iswcsymf_l⚠

_iswctype_l⚠

_iswdigit_l⚠

_iswgraph_l⚠

_iswlower_l⚠

_iswprint_l⚠

_iswpunct_l⚠

_iswspace_l⚠

_iswupper_l⚠

_iswxdigit_l⚠

_isxdigit_l⚠

_m_prefetchw⚠

_memccpy⚠

_memicmp⚠

_memicmp_l⚠

_mm_clflush⚠

_mm_getcsr⚠

_mm_lfence⚠

_mm_mfence⚠

_mm_pause⚠

_mm_prefetch⚠

_mm_setcsr⚠

_mm_sfence⚠

_mul128⚠

_rotl⚠

_rotl8⚠

_rotl16⚠

_rotl64⚠

_rotr⚠

_rotr8⚠

_rotr16⚠

_rotr64⚠

_strcmpi⚠

_strcoll_l⚠

_strdup⚠

_strerror⚠

_strerror_s⚠

_stricmp⚠

_stricmp_l⚠

_stricoll⚠

_stricoll_l⚠

_strlwr⚠

_strlwr_l⚠

_strlwr_s⚠

_strlwr_s_l⚠

_strncoll⚠

_strncoll_l⚠

_strnicmp⚠

_strnicmp_l⚠

_strnicoll⚠

_strnicoll_l⚠

_strnset⚠

_strnset_s⚠

_strrev⚠

_strset⚠

_strset_s⚠

_strupr⚠

_strupr_l⚠

_strupr_s⚠

_strupr_s_l⚠

_strxfrm_l⚠

_tolower⚠

_tolower_l⚠

_toupper⚠

_toupper_l⚠

_towlower_l⚠

_towupper_l⚠

_umul128⚠

_wcscoll_l⚠

_wcsdup⚠

_wcserror⚠

_wcserror_s⚠

_wcsicmp⚠

_wcsicmp_l⚠

_wcsicoll⚠

_wcsicoll_l⚠

_wcslwr⚠

_wcslwr_l⚠

_wcslwr_s⚠

_wcslwr_s_l⚠

_wcsncoll⚠

_wcsncoll_l⚠

_wcsnicmp⚠

_wcsnicmp_l⚠

_wcsnicoll⚠

_wcsnicoll_l⚠

_wcsnset⚠

_wcsnset_s⚠

_wcsrev⚠

_wcsset⚠

_wcsset_s⚠

_wcsupr⚠

_wcsupr_l⚠

_wcsupr_s⚠

_wcsupr_s_l⚠

_wcsxfrm_l⚠

is_wctype⚠

isalnum⚠

isalpha⚠

iscntrl⚠

isdigit⚠

isgraph⚠

isleadbyte⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

iswalnum⚠

iswalpha⚠

iswascii⚠

iswcntrl⚠

iswctype⚠

iswdigit⚠

iswgraph⚠

iswlower⚠

iswprint⚠

iswpunct⚠

iswspace⚠

iswupper⚠

iswxdigit⚠

isxdigit⚠

memccpy⚠

memchr⚠

memcmp⚠

memcpy⚠

memcpy_s⚠

memicmp⚠

memmove⚠

memmove_s⚠

memset⚠

strcat⚠

strcat_s⚠

strchr⚠

strcmp⚠

strcmpi⚠

strcoll⚠

strcpy⚠

strcpy_s⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_s⚠

stricmp⚠

strlen⚠

strlwr⚠

strncat⚠

strncat_s⚠

strncmp⚠

strncpy⚠

strncpy_s⚠

strnicmp⚠

strnlen⚠

strnset⚠

strpbrk⚠

strrchr⚠

strrev⚠

strset⚠

strspn⚠

strstr⚠

strtok⚠

strtok_s⚠

strupr⚠

strxfrm⚠

tolower⚠

toupper⚠

towlower⚠

towupper⚠

vDbgPrintEx⚠

vDbgPrintExWithPrefix⚠

wcscat⚠

wcscat_s⚠

wcschr⚠

wcscmp⚠

wcscoll⚠

wcscpy⚠

wcscpy_s⚠

wcscspn⚠

wcsdup⚠

wcsicmp⚠

wcsicoll⚠

wcslen⚠

wcslwr⚠

wcsncat⚠

wcsncat_s⚠

wcsncmp⚠

wcsncpy⚠

wcsncpy_s⚠

wcsnicmp⚠

wcsnlen⚠

wcsnset⚠

wcspbrk⚠

wcsrchr⚠

wcsrev⚠

wcsset⚠

wcsspn⚠

wcsstr⚠

wcstok⚠

wcstok_s⚠

wcsupr⚠

wcsxfrm⚠